requirements.txt
Django==4.2
environs==9.5.0
Pillow==9.5.0
drf-yasg==1.21.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-storages==1.13.2
django-storages[google]
django-cors-headers==4.0.0